requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
bcrypt==5.0.0
python-dotenv==1.2.1
requests==2.31.0
pandas
numpy
supabase==2.3.0
tensorflow
scikit-learn==1.3.0
yfinance==0.2.28
